Add probabilistic topic partition assignment
Thus far, the assumption of kafka-node's high-level consumer was that
there are more partitions than consumers in any given consumer group.
When that is not the case, the first consumer to subscribe gets all of
the listed partitions, which is not optimal. This commit introduces a
probabilistic partition assignment scheme which takes into account the
total number of subcribers when deciding whether to become the owner of
a partition.

In order to accommodate such behaviour, a Zookeeper-based locking
mechanism has been also implemented.