Check for partition owners in the consumer's group
When a consumer process isn't the owner of any partitions, it should
check that there are no partitions without an owner. If they do exist,
it should try to become their owner in a probabilistic manner. This
ensures that there are always consumers ready to receive messages for
topics even after client disconnects, crashes, etc.