Ignore NO_NODE errors when creating the consumer-topic map
On each rebalance and free partition check, the driver constructs the
consumer-topic and topic-consumer maps. However, as each consumer path
is checked separately, the consumer might have disappeared in the
meantime (possibly without ZK noticing it). As losing consumers is
acceptable, don't abort the whole process and leave it in an
undeterministic state when that happens.